refactor: improve types readibility
