+++ playground/pages/home.vue
@@ -57,7 +57,7 @@
 <script setup lang="ts">
 import { definePageMeta, useAuth, useAuthSession, ref, useAsyncData } from '#imports'
 
-definePageMeta({ middleware: 'auth' })
+definePageMeta({ middleware: 'auth',  })
 
 const { logout, fetchUser, changePassword } = useAuth()
 const { user, revokeAllSessions, getAllSessions, revokeSession } = useAuthSession()
